feat: extract our custom eleventy scss plugin
