+++ package.json
@@ -45,6 +45,7 @@
     "npm-run-all": "^4.1.5",
     "rimraf": "^6.0.1",
     "rss-parser": "^3.13.0",
+    "sass": "^1.87.0",
     "sass-embedded": "^1.87.0"
   },
   "overrides": {
